feat(static_pie): execute .init_array, while ignoring POSIX;
It looks as though I forgot to call the init array in static pie, and it looks as though there are some other important details:

```rs
/// glibc passes argc, argv, and envp to functions in .init_array, as a non-standard extension.
/// This allows `std::env::args` to work even in a `cdylib`, as it does on macOS and Windows.
```

These people really can't follow a standard... I do think this is a good idea though so I'll do the same.

One thing this commit fixs is now we can actually use rusts built in args...
But it means that if our target_env ever changes from gnu we're going to have other issues.